新增編輯視窗 && 新增表單欄位檢核與樣式 && 完成編輯功能

+++ src/app/pages/favorites/favorites.component.ts
@@ -1,7 +1,6 @@
 import { Component } from '@angular/core';
 
 import { TravelService } from 'libs/shared/src/lib/services/travel.service';
-import { BlockViewService } from 'libs/shared/src/lib/ui/block-view/block-view.service';
 import { BehaviorSubject, mergeMap, Observable, tap } from 'rxjs';
 
 @Component({
@@ -10,7 +9,6 @@ import { BehaviorSubject, mergeMap, Observable, tap } from 'rxjs';
   styleUrls: ['./favorites.component.scss'],
 })
 export class FavoritesComponent {
-
   selectedTravelList: any[] = [];
 
   update$ = new BehaviorSubject<any>('');
@@ -21,14 +19,21 @@ export class FavoritesComponent {
   );
 
   constructor(
-    private blockViewService: BlockViewService,
     public travelService: TravelService
   ) {}
 
   removeFavorites() {
+    if (this.selectedTravelList.length) {
+      return;
+    }
     this.travelService.removeFavorites(this.selectedTravelList);
     alert('移除成功');
     this.update$.next('');
   }
 
+  editTrave(data: any) {
+    this.travelService.modifyFavorite(data);
+    alert('修改成功');
+    this.update$.next('');
+  }
 }
